reorganize bundles
