+++ webpack.config.js
@@ -26,9 +26,9 @@ const config = {
     articleLandingPage: './cigionline/static/pages/article_landing_page/index.js',
     articlePage: './cigionline/static/pages/article_page/index.js',
     cigionline: './cigionline/static/index.js',
-    multimediaListPage: './cigionline/static/multimedia_list_page.js',
+    multimediaListPage: './cigionline/static/pages/multimedia_list_page/index.js',
     multimediaPage: './cigionline/static/pages/multimedia_page/index.js',
-    publicationListPage: './cigionline/static/publication_list_page.js',
+    publicationListPage: './cigionline/static/pages/publication_list_page/index.js',
     publicationPage: './cigionline/static/pages/publication_page/index.js',
   },
